join argument in sgs.get
The sgs.get function returns a list with dataframes

To have all dataframes joined into a single dataframe the argument join has to be set to True.

The default return is not joined by default because series can nave different sizes and, in these situations, a decision by the caller in necessary.

The join argument doen't solve that, but once the caller understands the situation, the join argument can be set to False.

Maybe in the future the join argument accepts str specifing what kind of join has to be executed.